Organising Project files Seperate buttons to new component
