--- src/components/Calculator.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import Container from './Container1';
-
-export default class Calculator extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-    };
-  }
-
-  render() {
-    return (
-      <Container />
-    );
-  }
-}
